+++ packages/commcare/src/Adaptor.js
@@ -1,16 +1,10 @@
-
-
-import {
-  execute as commonExecute,
-  http,
-  expandReferences,
-} from '@openfn/language-common';
-import pkg from '@openfn/language-http';
-const { get, post } = pkg
-import request from 'superagent';
-import FormData from 'form-data';
+import { execute as commonExecute } from '@openfn/language-common';
+import { expandReferences } from '@openfn/language-common/util';
+import { Blob } from 'node:buffer';
 import js2xmlparser from 'js2xmlparser';
 import xlsx from 'xlsx';
+
+import { request, prepareNextState } from './Utils';
 
 /**
  * Execute a sequence of operations.
@@ -31,36 +25,105 @@ export function execute(...operations) {
   };
 
   return state => {
-    state.configuration.authType = 'basic';
-    state.configuration.baseUrl = 'https://www.commcarehq.org/a/'.concat(
-      state.configuration.applicationName
-    );
     return commonExecute(...operations)({ ...initialState, ...state });
   };
 }
 
 /**
- * Performs a post request
- * @example
- *  clientPost(formData)
+ * Make a get request to any commcare endpoint
+ * - The response returned is {meta:{}, objects:[]}. These are destructured where objects will be written into state.data and meta into state.response along with the status code and returned headers.
+ * @public
+ * @example <caption>Get a list of cases</caption>
+ * get(
+ *    "case"
+ *    {
+ *      limit: 1,
+ *      offset:0,
+ *    }
+ * )
+ *  * @example <caption>Get a specific case </caption>
+ * get(
+ *    "case/12345"
+ *    {
+ *      limit: 1,
+ *      offset:0,
+ *    }
+ * )
  * @function
- * @param {Object} formData - Form Data with auth params and body
- * @returns {State}
+ * @param {string} path - Path to resource
+ * @param {Object} params - Optional request params such as limit and offset.
+ * @param {function} [callback] - Optional callback to handle the response
+ * @returns {Operation}
  */
-function clientPost({ url, body, username, password }) {
-  return new Promise((resolve, reject) => {
-    request
-      .post(url)
-      .auth(username, password)
-      .set('Content-Type', 'application/xml')
-      .send(body)
-      .end((error, res) => {
-        if (!!error || !res.ok) {
-          reject(error);
+export function get(path, params = {}, callback = s => s) {
+  return async state => {
+    const { domain } = state.configuration;
+    const [resolvedPath, resolvedParams] = expandReferences(
+      state,
+      path,
+      params
+    );
+
+    try {
+      const response = await request(
+        state.configuration,
+        `/a/${domain}/api/v0.5/${resolvedPath}`,
+        {
+          method: 'GET',
+          params: resolvedParams,
+          contentType: 'application/json',
         }
-        resolve(res);
-      });
-  });
+      );
+
+      return prepareNextState(state, response, callback);
+    } catch (e) {
+      throw e.body ?? e;
+    }
+  };
+}
+
+/**
+ * Make a post request to commcare
+ * @example
+ * post(
+ *   "user",
+ *  {"username":"test",
+ *  "password":"somepassword"}
+ * );
+ * @function
+ * @param {string} path - Path to resource
+ * @param {object} data - Object or JSON which defines data that will be used to create a given instance of resource
+ * @param {Object} params - Optional request params.
+ * @param {function} [callback] - Optional callback to handle the response
+ * @returns {Operation}
+ */
+export function post(path, data, params = {}, callback = s => s) {
+  return async state => {
+    const { domain } = state.configuration;
+    const [resolvedPath, resolvedData, resolvedParams] = expandReferences(
+      state,
+      path,
+      data,
+      params
+    );
+
+    try {
+      const response = await request(
+        state.configuration,
+        `/a/${domain}/api/v0.5/${resolvedPath}`,
+        {
+          method: 'POST',
+          data: resolvedData,
+          params: resolvedParams,
+          contentType: 'application/json',
+        }
+      );
+
+      return prepareNextState(state, response, callback);
+    } catch (e) {
+      throw e.body.error ?? e;
+    }
+  };
 }
 
 /**
@@ -83,17 +146,13 @@ function clientPost({ url, body, username, password }) {
  * @returns {Operation}
  */
 export function submitXls(formData, params) {
-  return state => {
-    const { applicationName, hostUrl, username, apiKey } = state.configuration;
+  return async state => {
+    const { domain } = state.configuration;
 
-    const json = expandReferences(formData)(state);
+    const [json] = expandReferences(state, formData);
     const { case_type, search_field, create_new_cases } = params;
 
-    const url = (hostUrl || 'https://www.commcarehq.org').concat(
-      '/a/',
-      applicationName,
-      '/importer/excel/bulk_upload_api/'
-    );
+    const path = `/a/${domain}/importer/excel/bulk_upload_api/`;
 
     const workbook = xlsx.utils.book_new();
     const worksheet = xlsx.utils.json_to_sheet(json);
@@ -106,28 +165,22 @@ export function submitXls(formData, params) {
 
     const data = new FormData();
 
-    data.append('file', buffer, { filename: 'output.xls' });
-    // data.append('file', fs.createReadStream('./out.xls'));
+    data.append('file', new Blob([buffer]), 'output.xls');
+    // data.append('file', fs.createReadStream('out.xls'));
     data.append('case_type', case_type);
     data.append('search_field', search_field);
     data.append('create_new_cases', create_new_cases);
 
-    console.log('Posting to url: '.concat(url));
-    return http
-      .post({
-        url,
+    try {
+      const response = await request(state.configuration, path, {
+        method: 'POST',
         data,
-        headers: {
-          ...data.getHeaders(),
-          Authorization: `ApiKey ${username}:${apiKey}`,
-        },
-      })(state)
-      .then(response => {
-        return { ...state, data: { body: response.data } };
-      })
-      .catch(err => {
-        throw { ...err, config: {}, request: {} };
       });
+
+      return prepareNextState(state, response);
+    } catch (e) {
+      throw e.body ?? e;
+    }
   };
 }
 
@@ -151,42 +204,25 @@ export function submitXls(formData, params) {
  * @returns {Operation}
  */
 export function submit(formData) {
-  return state => {
-    const jsonBody = expandReferences(formData)(state);
+  return async state => {
+    const [jsonBody] = expandReferences(state, formData);
     const body = js2xmlparser('data', jsonBody);
 
-    const {
-      // this should be called project URL.
-      // it is what lives after www.commcarehq.org/a/...
-      applicationName,
-      username,
-      password,
-      appId,
-      hostUrl,
-    } = state.configuration;
+    const { domain, appId } = state.configuration;
 
-    const url = (hostUrl || 'https://www.commcarehq.org').concat(
-      '/a/',
-      applicationName,
-      '/receiver/',
-      appId,
-      '/'
-    );
+    const path = `/a/${domain}/receiver/${appId}/`;
 
-    console.log('Posting to url: '.concat(url));
     console.log('Raw JSON body: '.concat(JSON.stringify(jsonBody)));
     console.log('X-form submission: '.concat(body));
 
-    return clientPost({
-      url,
-      body,
-      username,
-      password,
-    }).then(response => {
-      console.log(`Server repsonded with a ${response.status}:`);
-      console.log(response);
-      return { ...state, references: [response, ...state.references] };
+    const response = await request(state.configuration, path, {
+      method: 'POST',
+      data: body,
+      contentType: 'text/xml',
+      parseAs: 'text',
     });
+
+    return prepareNextState(state, response);
   };
 }
 
@@ -203,29 +239,28 @@ export function submit(formData) {
  * @returns {Operation}
  */
 export function fetchReportData(reportId, params, postUrl) {
-  return http.get(`api/v0.5/configurablereportdata/${reportId}/`, {
-    query: function (state) {
-      console.log(
-        'getting from url: '.concat(
-          state.configuration.baseUrl,
-          `/api/v0.5/configurablereportdata/${reportId}/`
-        )
-      );
-      console.log('with params: '.concat(params));
-      return params;
-    },
-    callback: function (state) {
-      var reportData = state.response.body;
-      return post(postUrl, {
-        body: reportData,
-      })(state).then(function (state) {
-        delete state.response;
-        console.log('fetchReportData succeeded.');
-        console.log('Posted to: '.concat(postUrl));
-        return state;
-      });
-    },
-  });
+  return async state => {
+    const path = `/a/${state.configuration.domain}/api/v0.5/configurablereportdata/${reportId}/`;
+
+    console.log('with params: '.concat(JSON.stringify(params)));
+
+    const { body: reportData } = await request(state.configuration, path, {
+      method: 'GET',
+      contentType: 'application/json',
+    });
+
+    const result = await request(state.configuration, postUrl, {
+      method: 'POST',
+      params,
+      data: reportData,
+      contentType: 'application/json',
+    });
+
+    delete result.response;
+    console.log('fetchReportData succeeded.');
+    console.log('Posted to: '.concat(postUrl));
+    return prepareNextState(state, {});
+  };
 }
 
 export {